+++ src/main/java/com/bootios/alone/domain/location/repository/LocationRepository.java
@@ -1,6 +1,6 @@
-package com.bootios.alone.domain.location.domain.repository;
+package com.bootios.alone.domain.location.repository;
 
-import com.bootios.alone.domain.location.domain.entity.Location;
+import com.bootios.alone.domain.location.entity.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LocationRepository extends JpaRepository<Location, Long> {}
