+++ src/main/java/com/bootios/alone/domain/location/service/LocationService.java
@@ -1,7 +1,7 @@
 package com.bootios.alone.domain.location.service;
 
-import com.bootios.alone.domain.location.domain.entity.Location;
-import com.bootios.alone.domain.location.domain.repository.LocationRepository;
+import com.bootios.alone.domain.location.entity.Location;
+import com.bootios.alone.domain.location.repository.LocationRepository;
 import com.bootios.alone.domain.location.dto.LocationCreateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
