+++ src/main/java/com/bootios/alone/domain/post/repository/PostRepository.java
@@ -1,7 +1,8 @@
-package com.bootios.alone.domain.post.domain.repository;
+package com.bootios.alone.domain.post.repository;
 
-import com.bootios.alone.domain.post.domain.entity.Post;
+import com.bootios.alone.domain.post.entity.Post;
 import java.util.Optional;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
